% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_in.R
\name{False_Discovery_Rate}
\alias{False_Discovery_Rate}
\title{False Discovery Rate}
\usage{
False_Discovery_Rate(X, y, beta_hat, cut = 0.5)
}
\arguments{
\item{X}{Matrix of predictors}

\item{y}{Binary response variable (0 or 1)}

\item{beta_hat}{Coefficient vector}

\item{cut}{Cut off value}
}
\value{
False Discovery Rate
}
\description{
Calculates the false discovery rate based on the confusion matrix.
}
