% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_in.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic Regression}
\usage{
logistic_regression(X, y)
}
\arguments{
\item{X}{Matrix of predictors}

\item{y}{Binary response variable (0 or 1)}
}
\value{
Coefficient vector beta
}
\description{
Function to perform logistic regression using numerical optimization.
}
