% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_in.R
\name{plot_logistic_curve}
\alias{plot_logistic_curve}
\title{Plot Fitted Logistic Curve}
\usage{
plot_logistic_curve(X, beta_hat, y)
}
\arguments{
\item{X}{Matrix of predictors}

\item{beta_hat}{Coefficient vector}

\item{y}{Binary response variable (0 or 1)}
}
\description{
Function to plot the fitted logistic curve.
}
